orchestrator/requirements.txt
google-adk>=1.0
langsmith>=0.2
httpx>=0.27
fastapi>=0.115
uvicorn>=0.30
python-dotenv>=1.0
pydantic>=2.0
pytest>=8.0
pytest-asyncio>=0.24
a2a-python>=0.3
